Commonize code between upload/download tests
