+++ rust-runtime/aws-smithy-runtime/tests/stalled_stream_upload.rs
@@ -3,27 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use aws_smithy_async::time::TimeSource;
-use aws_smithy_runtime::{assert_str_contains, test_util::capture_test_logs::capture_test_logs};
-use aws_smithy_types::error::display::DisplayErrorContext;
-use bytes::Bytes;
-use std::time::Duration;
-use tracing::info;
-
-/// No really, it's 42 bytes long... super neat
-const NEAT_DATA: Bytes = Bytes::from_static(b"some really neat data");
-
-/// Ticks time forward by the given duration, and logs the current time for debugging.
-macro_rules! tick {
-    ($ticker:ident, $duration:expr) => {
-        $ticker.tick($duration).await;
-        let now = $ticker
-            .now()
-            .duration_since(std::time::SystemTime::UNIX_EPOCH)
-            .unwrap();
-        tracing::info!("ticked {:?}, now at {:?}", $duration, now);
-    };
-}
+#[macro_use]
+mod stalled_stream_common;
+use stalled_stream_common::*;
 
 /// Scenario: Successful upload at a rate above the minimum throughput.
 /// Expected: MUST NOT timeout.
@@ -94,14 +76,7 @@ async fn upload_too_slow() {
         drop(body_sender);
     });
 
-    let err = result
-        .await
-        .expect("no panics")
-        .expect_err("should have timed out");
-    assert_str_contains!(
-        DisplayErrorContext(&err).to_string(),
-        "minimum throughput was specified at 1 B/s, but throughput of 0 B/s was observed"
-    );
+    expect_timeout(result.await.expect("no panics"));
 }
 
 /// Scenario: The server stops asking for data, the client maxes out its send buffer,
@@ -127,14 +102,7 @@ async fn upload_stalls() {
         time.tick(Duration::from_secs(1)).await;
     });
 
-    let err = result
-        .await
-        .expect("no panics")
-        .expect_err("should have timed out");
-    assert_str_contains!(
-        DisplayErrorContext(&err).to_string(),
-        "minimum throughput was specified at 1 B/s, but throughput of 0 B/s was observed"
-    );
+    expect_timeout(result.await.expect("no panics"));
 }
 
 // Scenario: The server stops asking for data, the client maxes out its send buffer,
@@ -194,88 +162,9 @@ async fn user_provides_data_too_slowly() {
     assert_eq!(200, result.await.unwrap().expect("success").as_u16());
 }
 
-use test_tools::*;
-mod test_tools {
-    use aws_smithy_async::test_util::tick_advance_sleep::{
-        tick_advance_time_and_sleep, TickAdvanceSleep, TickAdvanceTime,
-    };
-    use aws_smithy_async::time::TimeSource;
-    use aws_smithy_runtime::client::{
-        orchestrator::operation::Operation,
-        stalled_stream_protection::{
-            StalledStreamProtectionInterceptor, StalledStreamProtectionInterceptorKind,
-        },
-    };
-    use aws_smithy_runtime_api::{
-        client::{
-            http::{
-                HttpClient, HttpConnector, HttpConnectorFuture, HttpConnectorSettings,
-                SharedHttpConnector,
-            },
-            orchestrator::{HttpRequest, HttpResponse},
-            runtime_components::RuntimeComponents,
-            stalled_stream_protection::StalledStreamProtectionConfig,
-        },
-        http::StatusCode,
-        shared::IntoShared,
-    };
-    use aws_smithy_types::{body::SdkBody, timeout::TimeoutConfig};
-    use bytes::Bytes;
-    use http_body_0_4::Body;
-    use pin_utils::pin_mut;
-    use std::{
-        collections::VecDeque,
-        convert::Infallible,
-        future::poll_fn,
-        mem,
-        pin::Pin,
-        task::{Context, Poll},
-        time::Duration,
-    };
-    use tracing::instrument::Instrument;
-
-    #[derive(Debug)]
-    struct FakeServer(SharedHttpConnector);
-
-    impl HttpClient for FakeServer {
-        fn http_connector(
-            &self,
-            _settings: &HttpConnectorSettings,
-            _components: &RuntimeComponents,
-        ) -> SharedHttpConnector {
-            self.0.clone()
-        }
-    }
-
-    struct ChannelBody {
-        receiver: tokio::sync::mpsc::Receiver<Bytes>,
-    }
-    impl http_body_0_4::Body for ChannelBody {
-        type Data = Bytes;
-        type Error = Infallible;
-
-        fn poll_data(
-            mut self: Pin<&mut Self>,
-            cx: &mut Context<'_>,
-        ) -> Poll<Option<Result<Self::Data, Self::Error>>> {
-            match self.receiver.poll_recv(cx) {
-                Poll::Ready(value) => Poll::Ready(value.map(|v| Ok(v))),
-                Poll::Pending => Poll::Pending,
-            }
-        }
-
-        fn poll_trailers(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Result<Option<http::HeaderMap>, Self::Error>> {
-            unreachable!()
-        }
-    }
-
-    pub fn channel_body() -> (SdkBody, tokio::sync::mpsc::Sender<Bytes>) {
-        let (sender, receiver) = tokio::sync::mpsc::channel(1000);
-        (SdkBody::from_body_0_4(ChannelBody { receiver }), sender)
-    }
+use upload_test_tools::*;
+mod upload_test_tools {
+    use crate::stalled_stream_common::*;
 
     pub fn successful_response() -> HttpResponse {
         HttpResponse::try_from(
@@ -420,5 +309,13 @@ mod test_tools {
             Vec<u64>,
             time_sequence.into_iter().collect()
         )
+    }
+
+    pub fn expect_timeout(result: Result<StatusCode, SdkError<Infallible, Response<SdkBody>>>) {
+        let err = result.expect_err("should have timed out");
+        assert_str_contains!(
+            DisplayErrorContext(&err).to_string(),
+            "minimum throughput was specified at 1 B/s, but throughput of 0 B/s was observed"
+        );
     }
 }
